refactor: New folder for Drone Widget
