+++ src/components/DroneWidget/DroneWidgetTable.vue
